use new codec trait methods to manipulate sequences

+++ bio-seq/src/seq/index.rs
@@ -10,8 +10,8 @@ impl<A: Codec> Index<Range<usize>> for Seq<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: Range<usize>) -> &Self::Output {
-        let s = range.start * A::BITS;
-        let e = range.end * A::BITS;
+        let s = range.start * A::BITS as usize;
+        let e = range.end * A::BITS as usize;
         let bs = &self.bv[s..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -21,7 +21,7 @@ impl<A: Codec> Index<RangeTo<usize>> for Seq<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeTo<usize>) -> &Self::Output {
-        let e = range.end * A::BITS;
+        let e = range.end * A::BITS as usize;
         let bs = &self.bv[..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -31,7 +31,7 @@ impl<A: Codec> Index<RangeToInclusive<usize>> for Seq<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeToInclusive<usize>) -> &Self::Output {
-        let e = (range.end + 1) * A::BITS;
+        let e = (range.end + 1) * A::BITS as usize;
         let bs = &self.bv[..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -41,8 +41,8 @@ impl<A: Codec> Index<RangeInclusive<usize>> for Seq<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeInclusive<usize>) -> &Self::Output {
-        let s = range.start() * A::BITS;
-        let e = (range.end() + 1) * A::BITS;
+        let s = range.start() * A::BITS as usize;
+        let e = (range.end() + 1) * A::BITS as usize;
         let bs = &self.bv[s..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -52,7 +52,7 @@ impl<A: Codec> Index<RangeFrom<usize>> for Seq<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeFrom<usize>) -> &Self::Output {
-        let s = range.start * A::BITS;
+        let s = range.start * A::BITS as usize;
         let bs = &self.bv[s..] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -79,8 +79,8 @@ impl<A: Codec> Index<Range<usize>> for SeqSlice<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: Range<usize>) -> &Self::Output {
-        let s = range.start * A::BITS;
-        let e = range.end * A::BITS;
+        let s = range.start * A::BITS as usize;
+        let e = range.end * A::BITS as usize;
         let bs = &self.bs[s..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -90,7 +90,7 @@ impl<A: Codec> Index<RangeTo<usize>> for SeqSlice<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeTo<usize>) -> &Self::Output {
-        let e = range.end * A::BITS;
+        let e = range.end * A::BITS as usize;
         let bs = &self.bs[..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -100,7 +100,7 @@ impl<A: Codec> Index<RangeToInclusive<usize>> for SeqSlice<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeToInclusive<usize>) -> &Self::Output {
-        let e = (range.end + 1) * A::BITS;
+        let e = (range.end + 1) * A::BITS as usize;
         let bs = &self.bs[..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -110,8 +110,8 @@ impl<A: Codec> Index<RangeInclusive<usize>> for SeqSlice<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeInclusive<usize>) -> &Self::Output {
-        let s = range.start() * A::BITS;
-        let e = (range.end() + 1) * A::BITS;
+        let s = range.start() * A::BITS as usize;
+        let e = (range.end() + 1) * A::BITS as usize;
         let bs = &self.bs[s..e] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
@@ -121,7 +121,7 @@ impl<A: Codec> Index<RangeFrom<usize>> for SeqSlice<A> {
     type Output = SeqSlice<A>;
 
     fn index(&self, range: RangeFrom<usize>) -> &Self::Output {
-        let s = range.start * A::BITS;
+        let s = range.start * A::BITS as usize;
         let bs = &self.bs[s..] as *const Bs as *const SeqSlice<A>;
         unsafe { &*bs }
     }
